+++ pkg/reconciler/pubsub/pullsubscription/static/controller.go
@@ -14,8 +14,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-// TODO better package name?
-package k8s
+package static
 
 import (
 	"context"
