+++ pkg/reconciler/pubsub/pullsubscription/static/pullsubscription_test.go
@@ -14,7 +14,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package k8s
+package static
 
 import (
 	"context"
